fix: Model로 변환을 Controller 계층에서 수행하도록 변경한다.
